move treeReduce and treeAggregate to core
